DB-11957 align with QPT version of code
