--- utilities/src/main/java/com/splicemachine/qpt/Configuration.java
@@ -1,27 +1,0 @@
-/*
- * Copyright (c) 2012 - 2021 Splice Machine, Inc.
- *
- * This file is part of Splice Machine.
- * Splice Machine is free software: you can redistribute it and/or modify it under the terms of the
- * GNU Affero General Public License as published by the Free Software Foundation, either
- * version 3, or (at your option) any later version.
- * Splice Machine is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
- * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- * See the GNU Affero General Public License for more details.
- * You should have received a copy of the GNU Affero General Public License along with Splice Machine.
- * If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-package com.splicemachine.qpt;
-
-public class Configuration {
-    enum PrepMode {
-        NONE,    // don't prepare
-        WHOLE,   // prepare the entire statement as is
-        AUTO,    // prepare with auto-parameterization
-        SUBST;   // prepare with parameter substitution
-    };
-
-    static PrepMode prepare = PrepMode.NONE;
-}
