+++ utilities/src/main/java/com/splicemachine/qpt/SQLStatement.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012 - 2021 Splice Machine, Inc.
+ * Copyright (c) 2021 Splice Machine, Inc.
  *
  * This file is part of Splice Machine.
  * Splice Machine is free software: you can redistribute it and/or modify it under the terms of the
@@ -10,20 +10,29 @@
  * See the GNU Affero General Public License for more details.
  * You should have received a copy of the GNU Affero General Public License along with Splice Machine.
  * If not, see <http://www.gnu.org/licenses/>.
- *
  */
 
 package com.splicemachine.qpt;
 
 import com.splicemachine.qpt.SQLTokenizer.Token;
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 
 import java.io.*;
-import java.nio.charset.StandardCharsets;
+import java.nio.charset.Charset;
 
+@SuppressFBWarnings("EI_EXPOSE_REP2")
 public class SQLStatement {
 
     private SQLSignature signature;
     private Token[]         tokens;
+
+    enum PrepMode {
+        NONE,    // don't prepare
+        WHOLE,   // prepare the entire statement as is
+        AUTO,    // prepare with auto-parameterization
+        SUBST;   // prepare with parameter substitution
+    };
+    static PrepMode prepareMode = PrepMode.NONE;
 
     public SQLStatement(Token[] tokens) {
         this.tokens = tokens;
@@ -35,7 +44,7 @@ public class SQLStatement {
     }
 
     public String getSQL() {
-        switch (Configuration.prepare) {
+        switch (prepareMode) {
             case NONE:
             case WHOLE:
                 return signature.getSQL(tokens);
@@ -45,7 +54,7 @@ public class SQLStatement {
     }
 
     public Token[] getParams() {
-        switch (Configuration.prepare) {
+        switch (prepareMode) {
             case NONE:
             case WHOLE:
                 return null;
@@ -69,8 +78,9 @@ public class SQLStatement {
 
     public static SQLStatement getSqlStatement(String statement) throws IOException {
         if(statement == null) statement = "";
-        InputStream is = new ByteArrayInputStream(statement.getBytes(StandardCharsets.UTF_8));
-        try (BufferedReader in = new BufferedReader(new InputStreamReader(is, StandardCharsets.UTF_8))) {
+        String charset = Charset.defaultCharset().name();
+        InputStream is = new ByteArrayInputStream(statement.getBytes(charset));
+        try (BufferedReader in = new BufferedReader(new InputStreamReader(is, charset))) {
             SQLTokenizer lexer = new SQLTokenizer(in);
             return new SQLStatement(lexer.tokenize());
         }
