+++ utilities/src/test/java/com/splicemachine/qpt/PlanParserTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012 - 2021 Splice Machine, Inc.
+ * Copyright (c) 2021 Splice Machine, Inc.
  *
  * This file is part of Splice Machine.
  * Splice Machine is free software: you can redistribute it and/or modify it under the terms of the
@@ -10,7 +10,6 @@
  * See the GNU Affero General Public License for more details.
  * You should have received a copy of the GNU Affero General Public License along with Splice Machine.
  * If not, see <http://www.gnu.org/licenses/>.
- *
  */
 
 package com.splicemachine.qpt;
@@ -134,7 +133,7 @@ public class PlanParserTest {
     @Test
     public void testPlanParserCornerCases() throws ParseException {
         expectFails("A(x=0)\n" +
-                     "       -> B(x=0)\n");
+                "       -> B(x=0)\n");
 
         expectFails("   A(x=0)\n" +
                 "       -> B(x=0)\n");
